Add support for xcodebuild exportLocalizations warnings (#91)
The `xcodebuild -exportLocalizations` subcommand will warn when you have reused a localized string key with different values. This adds support for forwarding that warning through xcbeautify. I've decided to keep the original message as it is informative and already quite terse. I didn't use `formatCompleteWarning` because there is a `--- WARNING: ` prefix before the message.