Fix SwiftCompileCaptureGroup Bug (#306)
Fix a bug in the `SwiftCompileCaptureGroup` where it missed files with
escaped spaces. Add a unit test that considers the previously missed
output, and simplify the regex to more exhaustively consider filenames.
As part of this change, update the regex's capturing groups, and use
`lastPathComponent` to extract `filename` from the already captured
`filePath`.