Delete LinkerDuplicateSymbolsLocationCaptureGroup (#368)
As a prerequisite to #365, delete this overly-generic definition that
conflicts with `ShellCommandCaptureGroup`. Given there aren't tests
against this regular expression, it's unclear what output it actually
expects and how it differs from `LinkerDuplicateSymbolsCaptureGroup`.
Given this definition is formatted as `nil` output, we can safely delete
this type; deleting this type won't impact user-facing logs, so the only
known trade-off is a potential time regression.